pyproject.toml
[project]
name = "langgraph-crawler"
version = "0.1.0"
description = "A 6-node LangGraph pipeline: intent parsing, URL discovery, web crawling, source verification, MongoDB logging, and preprocessing."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=0.5.0",
    "crawl4ai>=0.8.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "langchain>=1.2.10",
    "langchain-community>=0.3.0",
    "langchain-tavily>=0.1",
    "langgraph>=1.0.9",
    "motor>=3.6.0",
    "pandas>=2.2.0",
    "playwright>=1.58.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.1",
    "replicate>=1.0.0",
    "streamlit>=1.40.0",
    "tavily-python>=0.7.21",
    "uvicorn>=0.32.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["crawler"]
